+++ src/themeUtils.ts
@@ -395,6 +395,10 @@ export class ThemeUtils {
       "dependency-preview": { vscode: "gear", hardis: "hammer-wrench.svg" },
       "dependency-error": { vscode: "error", hardis: "error.svg" },
       "dependency-local": { vscode: "gear", hardis: "hammer-wrench.svg" },
+      "mcp:start-sf-cli-mcp-server": {
+        vscode: "server-process",
+        hardis: "salesforce.svg",
+      },
       loading: { vscode: "loading~spin", hardis: "refresh.svg" },
     };
   }